split out boxes
